Fix error that shows centroids for objects that should not be tracked.
Even when just one class (with or without specified properties) was
tracked, centroids of other objects in the scene were shown.

The fix implied a little change to the CentroidTracker: now, in the
self.objects dict, the values are tuples of (centroid, rect). So in the
main program, when we iterate over the objects, we just grab the
bounding box (along with all the other data) of that object, and proceed
to show this information as needed (if the object meets the criterias).