Almost finish ability to use counters.
The functionality for the counters is almost finished. Currently, it
keeps a registry of the objectIDs that has been tracked. We could sum up
all the values that are True for a given counter, and that would be the
unique objects of that class that has been tracked.

TODO: Some refactoring of the code.