Finish counter funcionality.
This commit doesn't add new functionalities, just removes some unused
arguments from the argument parser.

The basic counter functionalities are ready, now we need to just glue it
with the alerts and conditions from the language, so the counters can
increment when there is an `incr` instruction in the program, or in a
`when` statement.